requirements.txt
numpy
pandas
scikit-learn
psycopg2
matplotlib
seaborn
plotly
python-dotenv
#-e .